+++ CHANGELOG.md
@@ -4,9 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-## `1.0.1-29-SNAPSHOT`
+## `1.1.1-33-SNAPSHOT`
 
 Release date `UNRELEASED`
+
+## `1.1.0-32`
+
+Release date `2023-03-20`
 
 - (feat): implement `hsts` response header
 - (feat): implement `referrer-policy` response header